Update LBJ attribute data.xlsx
Now that I'm analyzing items set aside from the debitage, I modified the config file to help keep track of what the original class assignment was, so that it is easier to track when I change it. I forgot to add points, and other sub classes that we need in the industry so several points are classified as bifaces.